+++ packages/utils/functions/fetch/baseAppFetch/index.ts
@@ -8,6 +8,7 @@ import {
 import { LOGOUT_EVENT } from '../../../constants/events'
 import { SERVICES_WITHOUT_TOKEN } from '../../../constants/fetch'
 import { TokenTypes } from '../../../constants/token'
+import { templateEnv } from '../../env'
 import { eventEmitter } from '../../events'
 import { getLanguage } from '../../language/getLanguage'
 import { buildQueryString } from '../../string'
@@ -74,7 +75,7 @@ export const baseAppFetch: BaseAppFetch = async (
     accessCookieName = ACCESS_COOKIE_NAME,
     refreshCookieName = REFRESH_COOKIE_NAME,
     languageCookieName = LANGUAGE_COOKIE_NAME,
-    baseUrl = process.env.NEXT_PUBLIC_API_BASE_URL,
+    baseUrl = templateEnv.NEXT_PUBLIC_API_BASE_URL,
     servicesWithoutToken = SERVICES_WITHOUT_TOKEN,
     params = {},
     decamelizeRequestBodyKeys = true,
@@ -88,7 +89,7 @@ export const baseAppFetch: BaseAppFetch = async (
 ) => {
   const url = `${baseUrl}${path}`
   const isAuthTokenRequired = !servicesWithoutToken.some((regex) => regex.test(path || ''))
-  const tokenType = process.env.NEXT_PUBLIC_TOKEN_TYPE as TokenTypes | undefined
+  const tokenType = templateEnv.NEXT_PUBLIC_TOKEN_TYPE as TokenTypes | undefined
 
   const fetchOptions: RequestOptions = {
     ...options,